Change money to number

+++ src/entity/deliveryBill.ts
@@ -28,7 +28,7 @@ export class DeliveryBill extends SoftDeleteEntity {
     @Column('timestamp with time zone', { nullable: true })
     public collectAt: Date;
 
-    @Column('money', { nullable: true })
+    @Column('float', { nullable: true })
     public collectValue: number;
 
     @Column('text')
